insights/pyproject.toml
[project]
name = "insights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "load-dotenv>=0.1.0",
    "requests>=2.32.5",
]
